requirements.txt
asgiref==3.8.1
Django==5.0.6
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
gunicorn==22.0.0
packaging==24.1
pillow==10.3.0
PyJWT==2.8.0
sqlparse==0.5.0
tzdata==2024.1